Galactic confusion was supposed to be interpolated between 'pixels';  it wasn't.  There was a half-pixel offset (difference between center of the pixel and upper left corner).  The return value is now a double rather than an integer;  the higher precision isn't entirely meaningful,  but avoids stair-step behavior as an object moves.

+++ orb_func.cpp
@@ -4757,7 +4757,7 @@ int metropolis_search( OBSERVE *obs, const int n_obs, double *orbit,
 #include "pl_cache.h"
 
 void update_environ_dot_dat( void);     /* mpc_obs.cpp */
-int galactic_confusion( const double ra, const double dec);
+double galactic_confusion( const double ra, const double dec);
 void pop_all_orbits( void);         /* orb_func2.cpp */
 char *find_numbered_mp_info( const int number);    /* mpc_obs.cpp */
 #if !defined( _WIN32) && !defined( __WATCOMC__)
